# openapi.yml
openapi: 3.0.3
info:
  title: Menu API
  version: 0.0.3
servers:
  - url: http://localhost:3000
tags:
  - name: Vendor
  - name: Menu
  - name: Menu Item
paths:
  /vendor:
    post:
      summary: Create a vendor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/vendor"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vendor"
      tags:
        - Vendor
  /vendor/{id}:
    get:
      summary: Find a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vendor"
        "404":
          description: Vendor not found
      tags:
        - Vendor
    delete:
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vendor"
        "404":
          description: Vendor not found
      tags:
      - Vendor
    patch:
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/vendor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vendor"
        "404":
          description: Vendor not found
      tags:
        - Vendor
  /menu:
    post:
      summary: Create a menu
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/menu"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menu"
      tags:
        - Menu
  /menu/{id}:
    get:
      summary: Find a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menu"
        "404":
          description: Menu not found
      tags:
        - Menu
    delete:
      summary: Delete a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menu"
        "404":
          description: Menu not found
      tags:
      - Menu
    patch:
      summary: Update a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/menu"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menu"
        "404":
          description: Menu not found
      tags:
        - Menu
  /menu/{menuId}/item:
    post:
      summary: Create a menu item
      parameters:
        - name: menuId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/menuItem"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menuItem"
      tags:
        - Menu Item
  /menu/{menuId}/item/{id}:
    get:
      summary: Find a menu item
      parameters:
        - name: menuId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menuItem"
        "404":
          description: Menu item not found
      tags:
        - Menu Item
    delete:
      summary: Delete a menu item
      parameters:
        - name: menuId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menuItem"
        "404":
          description: Menu item not found
      tags:
      - Menu Item
    patch:
      summary: Update a menu item
      parameters:
        - name: menuId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/menuItem"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/menuItem"
        "404":
          description: Menu item not found
      tags:
        - Menu Item
components:
  schemas:
    vendor:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          format: phone
        website:
          type: string
          format: website
        email:
          type: string
          format: email
        id:
          type: integer
    menu:
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
        vendorId:
          type: integer
    menuItem:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        id:
          type: integer
        menuId:
          type: integer